+++ src/game/Laptop/AIMMembers.cc
@@ -1753,7 +1753,7 @@ void DisplayTextForMercFaceVideoPopUp(const wchar_t* const pString)
 	swprintf(gsTalkingMercText, lengthof(gsTalkingMercText), L"\"%ls\"", pString);
 
 	//Set the minimum time for the dialogue text to be present
-	usAimMercSpeechDuration =  wcslen( gsTalkingMercText ) * AIM_TEXT_SPEECH_MODIFIER;
+	usAimMercSpeechDuration = static_cast<UINT16>(wcslen( gsTalkingMercText ) * AIM_TEXT_SPEECH_MODIFIER);
 
 	if( usAimMercSpeechDuration < MINIMUM_TALKING_TIME_FOR_MERC )
 		usAimMercSpeechDuration = MINIMUM_TALKING_TIME_FOR_MERC;
